refactor(query-insights): extract dedicated queryInsights sub-router (WI-4)
Create src/webviews/documentdb/collectionView/queryInsights/queryInsightsRouter.ts
and move Stage 1/2/3 + executeQueryInsightsAction procedures (and the
readQueryInsightsDebugFile helper) into it as a pure relocation — procedure
bodies are unchanged.

Mount the new sub-router under collectionView.queryInsights so the four
procedures are now reachable at collectionView.queryInsights.* in the tRPC
client (e.g. collectionView.queryInsights.getQueryInsightsStage3.query).
Per D12 / the package README convention, the streaming subscription
landing in WI-5 will live in a sibling queryInsightsEventsRouter.ts merged
into queryInsightsRouter — keeping 'things the webview calls' separate
from 'things the host pushes'.

Update all four webview call sites (QueryInsightsTab.tsx + CollectionView.tsx)
to use the new paths. Remove now-unused imports from collectionViewRouter.ts.

⚠️ Telemetry: rpc event paths for these procedures now include a
'queryInsights' segment (e.g. documentDB.rpc.query.collectionView.queryInsights.getQueryInsightsStage3 —
previously documentDB.rpc.query.collectionView.getQueryInsightsStage3).
Telemetry queries that hard-coded the old path must be updated.

WI-4 of the Stage 3 progressive-streaming plan.
